refactor(frontend): move sidebar item and group code
refactors the sidebar code so that item and item group into its own
focused files.
