+++ frontend/src/components/sidebar/Sidebar.scss
@@ -44,7 +44,7 @@ $text-muted: #667085;
 
   .sidebar-header {
     height: $sidebar-header-height;
-    border-bottom: 1px solid $sidebar-border;
+    border-bottom: 2px solid $sidebar-border;
   }
 
   .sidebar-brand-img {
