+++ frontend/src/i18n/langs/en.json
@@ -1535,66 +1535,6 @@
     "defaultMessage": "Search...",
     "description": "Placeholder for the search input of the SearchBox"
   },
-  "components.Sidebar.ChannelsLabel": {
-    "defaultMessage": "Channels"
-  },
-  "components.Sidebar.applications.applicationsLabel": {
-    "defaultMessage": "Applications"
-  },
-  "components.Sidebar.applications.campaignsLabel": {
-    "defaultMessage": "Campaigns"
-  },
-  "components.Sidebar.applications.containersLabel": {
-    "defaultMessage": "Containers"
-  },
-  "components.Sidebar.applications.deploymentsLabel": {
-    "defaultMessage": "Deployments"
-  },
-  "components.Sidebar.applications.imageCredentialsLabel": {
-    "defaultMessage": "Image Credentials"
-  },
-  "components.Sidebar.applications.networksLabel": {
-    "defaultMessage": "Networks"
-  },
-  "components.Sidebar.applications.volumesLabel": {
-    "defaultMessage": "Volumes"
-  },
-  "components.Sidebar.baseImageCollectionsLabel": {
-    "defaultMessage": "Base Image Collections"
-  },
-  "components.Sidebar.deviceGroupsLabel": {
-    "defaultMessage": "Groups"
-  },
-  "components.Sidebar.devicesLabel": {
-    "defaultMessage": "Devices"
-  },
-  "components.Sidebar.fileDownloadCampaignsLabel": {
-    "defaultMessage": "File Download Campaigns"
-  },
-  "components.Sidebar.filesManagementGroupLabel": {
-    "defaultMessage": "Files Management"
-  },
-  "components.Sidebar.hardwareTypesLabel": {
-    "defaultMessage": "Hardware Types"
-  },
-  "components.Sidebar.logoutLabel": {
-    "defaultMessage": "Logout"
-  },
-  "components.Sidebar.modelsGroupLabel": {
-    "defaultMessage": "Models"
-  },
-  "components.Sidebar.modelsLabel": {
-    "defaultMessage": "System Models"
-  },
-  "components.Sidebar.otaUpdatesGroupLabel": {
-    "defaultMessage": "OTA Updates"
-  },
-  "components.Sidebar.repositoriesLabel": {
-    "defaultMessage": "Repositories"
-  },
-  "components.Sidebar.updateCampaignsLabel": {
-    "defaultMessage": "Update Campaigns"
-  },
   "components.StorageTable.freeSpaceTitle": {
     "defaultMessage": "Free Space"
   },
@@ -1766,6 +1706,66 @@
   },
   "components.WiFiScanResultsTable.seenAtTitle": {
     "defaultMessage": "Seen at"
+  },
+  "components.sidebar.Sidebar.devicesLabel": {
+    "defaultMessage": "Devices"
+  },
+  "components.sidebar.SidebarFooter.logoutLabel": {
+    "defaultMessage": "Logout"
+  },
+  "components.sidebar.groups.AppsSidebarGroup.applicationsLabel": {
+    "defaultMessage": "Applications"
+  },
+  "components.sidebar.groups.AppsSidebarGroup.campaignsLabel": {
+    "defaultMessage": "Application Campaigns"
+  },
+  "components.sidebar.groups.AppsSidebarGroup.containersLabel": {
+    "defaultMessage": "Containers"
+  },
+  "components.sidebar.groups.AppsSidebarGroup.deploymentsLabel": {
+    "defaultMessage": "Deployments"
+  },
+  "components.sidebar.groups.AppsSidebarGroup.imageCredentialsLabel": {
+    "defaultMessage": "Image Credentials"
+  },
+  "components.sidebar.groups.AppsSidebarGroup.networksLabel": {
+    "defaultMessage": "Networks"
+  },
+  "components.sidebar.groups.AppsSidebarGroup.volumesLabel": {
+    "defaultMessage": "Volumes"
+  },
+  "components.sidebar.groups.FilesSidebarGroup.fileDownloadCampaignsLabel": {
+    "defaultMessage": "File Download Campaigns"
+  },
+  "components.sidebar.groups.FilesSidebarGroup.filesManagementGroupLabel": {
+    "defaultMessage": "Files Management"
+  },
+  "components.sidebar.groups.FilesSidebarGroup.repositoriesLabel": {
+    "defaultMessage": "Repositories"
+  },
+  "components.sidebar.groups.FleetSidebarGroup.ChannelsLabel": {
+    "defaultMessage": "Channels"
+  },
+  "components.sidebar.groups.FleetSidebarGroup.deviceGroupsLabel": {
+    "defaultMessage": "Groups"
+  },
+  "components.sidebar.groups.FleetSidebarGroup.fleetGroupLabel": {
+    "defaultMessage": "Fleet Management"
+  },
+  "components.sidebar.groups.FleetSidebarGroup.hardwareTypesLabel": {
+    "defaultMessage": "Hardware Types"
+  },
+  "components.sidebar.groups.FleetSidebarGroup.modelsLabel": {
+    "defaultMessage": "System Models"
+  },
+  "components.sidebar.groups.OtaSidebarGroup.baseImageCollectionsLabel": {
+    "defaultMessage": "Base Image Collections"
+  },
+  "components.sidebar.groups.OtaSidebarGroup.otaUpdatesGroupLabel": {
+    "defaultMessage": "OTA Updates"
+  },
+  "components.sidebar.groups.OtaSidebarGroup.updateCampaignsLabel": {
+    "defaultMessage": "Update Campaigns"
   },
   "forms.CreateApplication.descriptionLabel": {
     "defaultMessage": "Description"
